feat: Add revalidatePath to getServices and getUsers
This commit adds the revalidatePath function from the next/cache package to the getServices and getUsers functions in the get-services.ts and get-users.ts files respectively. The revalidatePath function is called with the appropriate path to trigger a revalidation of the cache for the specified route. This ensures that the cache is updated when new services or users are added, improving the accuracy of the data displayed.